Initial commit.
This project is gonna be used in conjunction with my "daybooks". I use
daily notion pages. One thing that I don't have a very good version of
yet is linking to specific daybooks relative to today's date. The
intention of this project is to rebuild on a daily cadence and have
these links up to date.