Add some assertions for the tests.
Previously the tests would run the build, but it would make any
assertions. This commit introduces some assertions by mocking the file
system. It also make the test double for the notion client to be more
strict about fixtures existing.

The contents of what is generated is now more visible.