Fix the dates being wrong.
Notion appears to have changed the results of their API since I first
wrote this to return fuzzier matches and not an exact match via the
search method. It appeared to have been returning things in recently
touched order.

This commit changes to fetching a full page (which it was doing before)
but now checking the titles of the results and matching on that.