Add the generation of a mobile page.
I want to be able to open this within android. This commit generates a
separate redirect page for that that uses the `notion://` scheme which
should work based on this:
https://thomasjfrank.com/how-to-share-notion-links-that-open-directly-in-the-app/

Another option that I considered was using JavaScript and a single page.
That would kinda be practice for a thing we're doing at work.